requirements.txt
Django==3.0.3
django-environ==0.4.5
psycopg2==2.8.4
djangorestframework==3.11.0
django-filter==2.2.0

django-allauth==0.38.0

django-rest-auth==0.9.3
oauthlib==2.0.3
requests-oauthlib==0.8.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
